// .PRD/erp-refactor/v3/types/process-statement-of-accounts.ts
import { z } from 'zod';

export const ProcessStatementOfAccountsSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  report: z.enum(['General Ledger', 'Accounts Receivable']),
  from_date: z.string().optional(),
  posting_date: z.string().optional().default('Today'),
  company: z.string(),
  account: z.string().optional(),
  categorize_by: z.enum(['Categorize by Voucher', 'Categorize by Voucher (Consolidated)']).optional().default('Categorize by Voucher (Consolidated)'),
  cost_center: z.array(z.unknown()).optional(),
  territory: z.string().optional(),
  ignore_exchange_rate_revaluation_journals: z.boolean().optional().default(false),
  ignore_cr_dr_notes: z.boolean().optional().default(false),
  to_date: z.string().optional(),
  finance_book: z.string().optional(),
  currency: z.string().optional(),
  project: z.array(z.unknown()).optional(),
  payment_terms_template: z.string().optional(),
  sales_partner: z.string().optional(),
  sales_person: z.string().optional(),
  show_remarks: z.boolean().optional().default(false),
  based_on_payment_terms: z.boolean().optional().default(false),
  show_future_payments: z.boolean().optional().default(false),
  customer_collection: z.enum(['Customer Group', 'Territory', 'Sales Partner', 'Sales Person']).optional(),
  collection_name: z.string().optional(),
  primary_mandatory: z.boolean().optional().default(true),
  show_net_values_in_party_account: z.boolean().optional().default(false),
  customers: z.array(z.unknown()),
  print_format: z.string().optional(),
  orientation: z.enum(['Landscape', 'Portrait']).optional(),
  include_break: z.boolean().optional().default(true),
  include_ageing: z.boolean().optional().default(false),
  ageing_based_on: z.enum(['Due Date', 'Posting Date']).optional().default('Due Date'),
  letter_head: z.string().optional(),
  terms_and_conditions: z.string().optional(),
  enable_auto_email: z.boolean().optional().default(false),
  sender: z.string().optional(),
  frequency: z.enum(['Daily', 'Weekly', 'Biweekly', 'Monthly', 'Quarterly']).optional(),
  filter_duration: z.number().int().optional().default(1),
  start_date: z.string().optional().default('Today'),
  pdf_name: z.string().optional(),
  subject: z.string().optional(),
  cc_to: z.array(z.unknown()).optional(),
  body: z.string().optional(),
  help_text: z.string().optional(),
});

export type ProcessStatementOfAccounts = z.infer<typeof ProcessStatementOfAccountsSchema>;

export const ProcessStatementOfAccountsInsertSchema = ProcessStatementOfAccountsSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type ProcessStatementOfAccountsInsert = z.infer<typeof ProcessStatementOfAccountsInsertSchema>;